install

+++ install/setup.sh
@@ -2,7 +2,9 @@
 set -e
 mkdir -p /src/makaronLab/build/editor
 mv mLabEditor /src/makaronLab/build/editor/
-mv makaronLab.desktop ~/Desktop/
+if [ -e ~/Desktop/ ]; then
+    mv makaronLab.desktop ~/Desktop/
+fi
 mv plugins /src/makaronLab/build/plugins
 mv lib /src/makaronLab/build/lib
 mv packages /src/makaronLab/packages
